send fewer packets by combining writes when possible
This makes the non-lazy negotiator act *slightly* more like the lazy one.
Instead of sending the multistream header and *then* the protocol header, it
sends them both in the same packet.

This *shouldn't* cause any issues but we should still think carefully about
whether or not this is worth it. It:

1. Sends fewer packets.
2. Saves half an RTT.